Create HomePage page component
